requirements.txt
asgiref==3.8.1
Django==5.0.7
gunicorn==22.0.0
packaging==24.1
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
